S509089 xen_cpu_usage: fix live migration bug
It seems that with live migration between wheezy dom0s, the migrated
domU appears at the target dom0 with the current amount of seconds (cpu
seconds used on the source dom0) listed. As soon as migration finishes,
it drops to 0. This causes spikes of > 10000 % in the graphs.

Fix for now: ignore domU if it's not seen before and if state is ------

With newer xen in Jessie, incoming domU shows up with --incoming
appended to the name, so just also ignore those.

And refactor a bit. More filtering in python instead of external
commands.

State file contents are incompatible with previous ones, no automated
migration, I'm just gonna throw the current ones away if I roll out the
new plugin. :o)