mxruntime_: Monkeyfix v3 changing primary groups
We're changing the primary group of the trX users in Mendix Cloud v3
from the "users" group to their own group with the same name as the
username.

In order to read /proc/<pid>/smaps, the munin plugin primary gid needs
to be the same as the running JVM process.

So, here's kind of a wrapper that starts as root, then checks what the
actual user and groups are that it should run as, and then drops
privileges to them.

Note: If you're not Mendix Cloud v3, then you shouldn't be using this
plugin code. Use the mxruntime_ from m2ee-tools itself instead!