appdbnode_df: don't use mountpount as munin label
During storage migration, we will convert really old filesystems that
have the postgresql database on /var into using the more specific
/var/lib/postgresql, which we have been using for quite a while instead
now.

When this happens, we should not have _var as part of the munin rrd
file, since the whole graph history will be lost and it will start over
with an rrd file with _var_lib_postgreqsql in the name.

So, use something different instead as field name.