appdbnode_df: work around single disk rackspace nodes
So, in the end we want to also have the new graphs shown for legacy
rackspace nodes that are still in use. I prefer this over doing hacks to
hot-rewire graph urls in nginx on the trends server.

But, most of them have a single disk, instead of separate partition for
files and database storage. Since only showing 'OS files' is not very
nice for the end user, let's monkey patch the description of the single
disk if we detect this scenario, kthxbye.

Example scenario with separate database partition:

    rootfs.label OS files (max 2.89GiB)
    rootfs.draw LINE1
    rootfs.info Disk usage for OS files in bytes
    database.label Database storage (max 4.80GiB)
    database.draw LINE1
    database.info Disk usage for Database storage in bytes

Example rackspace scenario with single disk partition on a dbnode:

    rootfs.label OS and Database files (max 19.68GiB)
    rootfs.draw LINE1
    rootfs.info Disk usage for OS and Database files in bytes